pyproject.toml
[tool.poetry]
name = "sajuai"
version = "0.1.0"
description = ""
authors = ["Madung2 <tuliphan91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
xmltodict = "^0.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
